Refactoring second user story
